+++ Win32/Sources/Application/Calendar/Calendar_Properties/CPropCalendarGeneral.cpp
@@ -137,13 +137,13 @@ void CPropCalendarGeneral::SetCalendar(calstore::CCalendarStoreNode* node)
 	if (!node->IsDirectory())
 	{
 		{
-			unsigned long size = node->GetSize();
-			if ((size == ULONG_MAX) && (node->GetProtocol()->IsOffline() || node->GetProtocol()->IsWebCalendar()))
+			uint32_t size = node->GetSize();
+			if ((size == UINT32_MAX) && (node->GetProtocol()->IsOffline() || node->GetProtocol()->IsWebCalendar()))
 			{
 				node->CheckSize();
 				size = node->GetSize();
 			}
-			if (size != ULONG_MAX)
+			if (size != UINT32_MAX)
 			{
 				mSize = ::GetNumericFormat(size);
 			}
@@ -223,7 +223,7 @@ void CPropCalendarGeneral::OnCalculateSize()
 
 	{
 		uint32_t size = mCalList->front()->GetSize();
-		if (size != ULONG_MAX)
+		if (size != UINT32_MAX)
 		{
 			CUnicodeUtils::SetWindowTextUTF8(GetDlgItem(IDC_CALENDARGENERAL_SIZE), ::GetNumericFormat(size));
 		}
